Add Jasmine test framework
Jasmine tests are located under src/tests/.  Two tasks have been added
to the gulpfile for building and running tests.

  test:
    Run all tests in dist/tests/.

  watch:
    When a file changes, run all tests in dist/tests/.

The 'watch' test has been changed so that it does not itself build all
files when it starts.  Instead it delegates to the 'build' task for an
initial build and only re-builds files that change.

Closes #2.